final cleaned up , should be all good
