+++ config/good_config/test.conf
@@ -41,6 +41,7 @@ server {
 
 
  location  /upload {
+    client_max_body_size 12k ;
         cgi .php welcome.php;        
         set_method POST DELETE GET on;
         upload_location /upload/;
@@ -50,4 +51,4 @@ server {
 		autoindex on;
         upload_location /upload/;
 	}
-    }
+	}
