+++ html/lim/uploadTest.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>File Upload Form</title>
+</head>
+
+<body>
+
+    <form action="/upload" method="post" enctype="multipart/form-data">
+        <input type="file" name="file1" id="file1" multiple>
+        <input type="file" name="file2" id="file2" multiple>
+        <input type="submit" value="Upload">
+    </form>
+</body>
+
+</html>
