--- html/upload/418.html
@@ -1,86 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>418 I'm a teapot at /html/ on exact match</title>
-    <style>
-        body {
-            background-color: #f8f8f8;
-            color: #333;
-            font-family: 'Arial', sans-serif;
-            text-align: center;
-            padding: 50px;
-        }
-
-        .container {
-            max-width: 600px;
-            margin: 0 auto;
-        }
-
-        h1 {
-            font-size: 120px;
-            margin: 0;
-        }
-
-        h2 {
-            font-size: 40px;
-            margin: 0;
-        }
-
-        p {
-            font-size: 18px;
-        }
-
-        a {
-            color: #007BFF;
-            text-decoration: none;
-        }
-
-        a:hover {
-            text-decoration: underline;
-        }
-    </style>
-</head>
-
-<body>
-    <div class="container">
-        <h1>418</h1>
-        <h2>I'm a teapot on /html/ and exact match testing a larger file count for the multipart please make it work:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa</h2>
-        <h3>gnfdiggferighnghighfjghjgkjfgshfwiuhrw vqoirhqworoqwjroq\r\npqkposjrorhioahreashreawrhjashrjshohjsahruwhruwu qworhqoejp0qwj0qjrpjfrdsjfldksjf argndfkignf</h3>
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-
-        <h4>grekjngurigrioejhfiorhgiofhgifghdighiufdhfsiudhgiudshfoidsjfgosdjgfhosdijfsojfsoijfois\r\ngeuuhfeiuwhfiewhfi\r\nfheifheufhwieofheowifhwe\r\nfewinfowfowfosa\r\n</h4>    
-        <p>Sorry, but you do not have permission to access this page.</p>
-        <p><a href="/">Go to Homepage</a></p>
-    </div>
-</body>
-
-</html>
