--- log/parsing_error.log
@@ -1,1 +1,0 @@
-
