--- log/server_error.log
@@ -1,6 +1,0 @@
-ip: 127.0.0.1; fd: 5; Port: 4253; 09-08-24 23:43:46 GMT; /favicon.ico; GET /favicon.ico; 301  Moved Permanently
-ip: 127.0.0.1; fd: 5; Port: 4253; 09-08-24 23:44:12 GMT; /upload; POST /upload; 201  Created
-ip: 127.0.0.1; fd: 5; Port: 4253; 09-08-24 23:44:12 GMT; /favicon.ico; GET /favicon.ico; 301  Moved Permanently
-ip: 127.0.0.1; fd: 5; Port: 4253; 09-08-24 23:44:33 GMT; /favicon.ico; GET /favicon.ico; 301  Moved Permanently
-ip: 127.0.0.1; fd: 5; Port: 4253; 09-08-24 23:44:51 GMT; /upload; POST /upload; 201  Created
-ip: 127.0.0.1; fd: 5; Port: 4253; 09-08-24 23:44:51 GMT; /favicon.ico; GET /favicon.ico; 301  Moved Permanently
