--- log/stderr.log
@@ -1,6 +1,0 @@
-Client::parseRequest: bool _isCgi: 0
-Client::parseRequest: bool _isCgi: 0
-Client::parseRequest: bool _isCgi: 0
-Client::parseRequest: bool _isCgi: 0
-Client::parseRequest: bool _isCgi: 0
-webserv: Webserv::start: epoll_wait: Interrupted system call
