--- srcs/request_parsing/OpenDir.cpp
@@ -1,1 +1,0 @@
-
