--- srcs/request_parsing/note.txt
@@ -1,45 +1,0 @@
-request parsing:
-
-skip first empty lines only first empty lines
-
-request: Error:   -strcmp != 0 with GET/POST/DELETE   => 400 Bad Request
-                  -method not allowed in this PATH    => 405 Not Allowed
-                  -if HTTP/1.0                        => dont keep connection alive
-                  -if HTTP/ missing or not 1.0.1.2.3  => 400 Bad Request
-                  -incorrect method                   => 405
-                  -if location doesnt start with /    => 400
-                  -location missing                   => 400
-                  -HTTP version needed, not "/" or nb => 400
-                  -HTTP version >= 2                  => 505 HTTP VERSION NOT SUPPORTED
-                  -HTTP version between 1 and 2(ni)   => OK return HTTP 1.1
-
-
-
-
-
-host: Error:  -field empty/missing                    => 400
-              -contains "/"                           => 400
-              -multiple host                          => 400 Bad Request 
-              -if wrong host                          => sent to default server ?
-
-userAgent: osef
-
-Accept: osef
-
-ignore useless info 
-
-!! missing empty_line at the end = 0 request
-
-RESPONSE:
-
-HTTP/1.1 STATUS_CODE(200 if ok) short description (OK)
-Server: servername
-Date: Mon, 24 Jun 2024 14:03:15 GMT
-Content-Type: text/html
-Content-Lenght: sizeofbody in bytes
-Last-Modified: date
-Connection: keep-alive if HTTP/1.1, close if HTTP/1.0
-ETag: "66771347-86" cache_opti
-Accept-Ranges: bytes ?
-EMPTY LINE
-BODY
